brown_belt: added solutions for priority collection challenge.

+++ belts_cpp/04_brown_belt/01_week/00_experiments/02_count_words_in_sherlock_holmes/profile.h
@@ -42,7 +42,7 @@ class LogDurationCallback : public LogDuration {
         : LogDuration("", true), callback_time(callable) {}
 
     ~LogDurationCallback() {
-        callback_time = GetCurrentTime();
+        callback_time = static_cast<double>(GetCurrentTime());
     }
 };
